fix create link

+++ create_link.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+read -p "this operation will delete all files in ./include/nagisa, ok?[y/N] " ok
+if [[ "$ok" != "y" && "$ok" != "Y" ]]; then
+    exit 0
+fi
+
 rm -rf ./include/nagisa
 mkdir ./include/nagisa
-python ./tools/create_link/main.py ./include/nagisa ./submodules
+python ./tools/create_link/main.py ./include/nagisa ./libs
